+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.2
+
+Features:
+
+- added working status (visual only, doesn't work yet)
+- added Pause and Skip button (visual only, doesn't work yet)
+- added Cycles Statistics (visual only, doesn't work yet)
+
+
 ## 0.1.1
 
 Features: